requirements.txt
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
requests==2.31.0
google-genai>=0.1.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0